pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr-optimizer"
version = "0.1.0"
description = "Multi-pass OCR optimizer using GLM-OCR via Ollama"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
ocr-optimize = "ocr_optimizer.cli:main"
